% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_mle.R
\name{params_mle}
\alias{params_mle}
\title{params_mle}
\usage{
params_mle(data)
}
\arguments{
\item{data, }{DATAFRAME, ARRAY, LISTS or VECTORS where data for each continous variable is in its respective column/list}
}
\value{
DATAFRAME where the first row contains the estimated means and the second row contains the estimated variance, and the columns present the original variables in the data
}
\description{
Fit data to a Guassian distribution with Maximum Likelihood Estimation (MLE)
}
\examples{
iris_data <- data.frame("length" = c(1,2,3,4), "width" = c(5,6,7,8))
params_mle(iris_data)

}
